+++ CHANGELOG.md
@@ -3,9 +3,10 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Umreleased
+## Unreleased
 ### Added
 - Report `textarea` and `select` elements.
+- Report interactable state of elements.
 
 ## 0.1.9 - 2026-05-15